Version 1.1
Code to attach to the PowerShell host is now called by the Add-LogFile
and Add-OutputSubscriber cmdlets, instead of by the psm1 file at module
load time.  This is to address an issue where the module auto-loading
due to tab completion was somehow causing it not to attach to the host
properly.